requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
openai==1.2.4
python-dotenv==1.0.0
chromadb==0.4.15
requests==2.31.0
beautifulsoup4==4.12.2
pydantic==2.4.2